requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
Pillow==10.1.0
python-decouple==3.8
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
coverage==7.3.2
pytest==7.4.3
pytest-django==4.7.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0 